Tride to do something with the footer; made a base situation for the company page
