+++ pages/components/FooterComponents/Footer.js
@@ -6,37 +6,33 @@ import LogoSVG from '../MenuComponents/Logo.svg';
 
 const FooterPolygon = styled(FooterPolygonSVG)`
   fill: ${props => props.theme.orange};
+  position: absolute;
+  z-index: 0;
+  bottom: 0;
 `;
 
 const RightPolygon = styled(RightPolygonSVG)`
   fill: ${props => props.theme.secondary};
+  position: absolute;
+  z-index: 1;
+  bottom: 0;
 `;
 
 const Logo = styled(LogoSVG)`
   fill: ${props => props.theme.primary};
-  position: fixed;
-  right: 5px;
-  bottom: 0px;
-  width: 30%;
-`
+  position: absolute;
+`;
 
 const Footer = ({ className }) => {
   return (
     <div className={className}>
       <FooterPolygon />
       <RightPolygon />
-      <Logo />
     </div>
   );
 };
 
 const StyledFooter = styled(Footer)`
-  svg {
-    position: fixed;
-    bottom: 0;
-  }
-
   width: 100%;
-  height: 100%;
 `;
 export default StyledFooter;
