+++ pages/components/MenuComponents/MenuTab.js
@@ -3,7 +3,7 @@ import { css } from 'styled-components';
 import styled from 'styled-components';
 import Link from 'next/link';
 
-const MenuTab = ({ className, link, text, browserUrl }) => {
+const MenuTabContainer = ({ className, link, text, browserUrl }) => {
   return (
     <Link href={link} as={browserUrl}>
       <a className={className}>{text}</a>
@@ -11,25 +11,10 @@ const MenuTab = ({ className, link, text, browserUrl }) => {
   );
 };
 
-const StyledMenuTab = styled(MenuTab)`
-  font-style: normal;
-  font-weight: 900;
-  font-size: 1.5rem;
-  line-height: 2em;
+const MenuTab = styled(MenuTabContainer)`
   text-decoration: none;
   text-align: center;
   color: ${props => props.theme.secondary};
-
-  @media only screen and (max-width: 768px) {
-    font-size: 1.5em;
-  }
   opacity: ${props => (props.isActive ? 1 : 0.7)};
-
-  ${props =>
-    props.underline &&
-    css`
-      border-bottom: ${props.isActive ? '3px' : '2px'} solid;
-      border-bottom-color: ${props => (props.isActive ? props.theme.orange : '#c4c4c4')};
-    `}
 `;
-export default StyledMenuTab;
+export default MenuTab;
